+++ deno.json
@@ -2,12 +2,11 @@
   "name": "@fcrozatier/monarch",
   "version": "2.6.0",
   "exports": {
-    ".": "./mod.ts",
+    ".": "./src/mod.ts",
     "./alternation": "./src/combinators/alternation/mod.ts",
     "./iteration": "./src/combinators/iteration/mod.ts",
     "./sequencing": "./src/combinators/sequencing/mod.ts",
     "./lazy": "./src/combinators/lazy-evaluation/mod.ts",
-    "./primitives": "./src/primitives/mod.ts",
     "./common": "./src/common/mod.ts",
     "./examples": "./examples/mod.ts"
   },
@@ -18,6 +17,7 @@
   "imports": {
     "$combinators": "./src/combinators/mod.ts",
     "$common": "./src/common/mod.ts",
+    "$core": "./src/mod.ts",
     "@std/assert": "jsr:@std/assert@^1"
   },
   "publish": {
