+++ mod.ts
@@ -1,3 +1,2 @@
 export * from "$combinators";
-
-export * from "./src/primitives/mod.ts";
+export * from "$core";
