+++ src/combinators/alternation/any.test.ts
@@ -1,10 +1,34 @@
+import { many, repeat } from "$combinators";
+import { literal, take } from "$common";
+import { parseErrors, type Parser } from "$core";
 import { assertEquals } from "@std/assert";
 import { any } from "./any.ts";
-import { parseErrors } from "../../errors.ts";
-import { take, takeTwo } from "../../primitives/mod.ts";
-import { literal } from "../../common/mod.ts";
+
+/**
+ * Parses the next two characters
+ */
+export const takeTwo: Parser<string> = repeat(take, 2).map((arr) =>
+  arr.join("")
+).error(parseErrors.takeTwoError);
 
 const oneOrTwoItems = any(take, takeTwo);
+
+Deno.test("take two", () => {
+  assertEquals(takeTwo.parse("m"), {
+    success: false,
+    message: parseErrors.takeTwoError,
+    position: { line: 1, column: 0 },
+  });
+
+  assertEquals(takeTwo.parse("monad"), {
+    success: true,
+    results: [{
+      value: "mo",
+      remaining: "nad",
+      position: { line: 1, column: 2 },
+    }],
+  });
+});
 
 Deno.test("any", () => {
   assertEquals(oneOrTwoItems.parse(""), {
@@ -37,5 +61,37 @@ Deno.test("any", () => {
     success: false,
     message: "",
     position: { line: 1, column: 1 },
+  });
+});
+
+// Explore a search space
+const explore = many(oneOrTwoItems);
+
+Deno.test("explore", () => {
+  assertEquals(explore.parse("many"), {
+    success: true,
+    results: [
+      {
+        remaining: "",
+        value: ["m", "a", "n", "y"],
+        position: { line: 1, column: 4 },
+      },
+      {
+        remaining: "",
+        value: ["m", "a", "ny"],
+        position: { line: 1, column: 4 },
+      },
+      {
+        remaining: "",
+        value: ["m", "an", "y"],
+        position: { line: 1, column: 4 },
+      },
+      {
+        remaining: "",
+        value: ["ma", "n", "y"],
+        position: { line: 1, column: 4 },
+      },
+      { remaining: "", value: ["ma", "ny"], position: { line: 1, column: 4 } },
+    ],
   });
 });
