+++ src/combinators/iteration/many1.ts
@@ -1,6 +1,6 @@
 import { many } from "./many.ts";
-import type { Parser } from "../../index.ts";
-import { result } from "../../primitives/result.ts";
+import type { Parser } from "$core";
+import { result } from "$core";
 
 /**
  * Returns the longest matching parse array (1 or more matches)
