+++ src/combinators/iteration/sepBy.ts
@@ -1,6 +1,6 @@
 import { alt } from "../alternation/alt.ts";
-import type { Parser } from "../../index.ts";
-import { result } from "../../primitives/result.ts";
+import type { Parser } from "$core";
+import { result } from "$core";
 import { sepBy1 } from "./sepBy1.ts";
 
 /**
