+++ src/combinators/iteration/sepBy1.ts
@@ -1,5 +1,5 @@
-import type { Parser } from "../../index.ts";
-import { result } from "../../primitives/result.ts";
+import type { Parser } from "$core";
+import { result } from "$core";
 import { many } from "./many.ts";
 
 /**
