+++ src/combinators/sequencing/foldL1.ts
@@ -1,6 +1,6 @@
 import { alt } from "../alternation/mod.ts";
-import type { Parser } from "../../index.ts";
-import { result } from "../../primitives/mod.ts";
+import type { Parser } from "$core";
+import { result } from "$core";
 
 /**
  * Parses non-empty sequences of items separated by an operator parser that associates to the left and performs the fold
