+++ src/common/characters.test.ts
@@ -1,6 +1,23 @@
+import { letter, lower, take, upper } from "$common";
 import { assertEquals } from "@std/assert";
 import { parseErrors } from "../errors.ts";
-import { letter, lower, upper } from "./mod.ts";
+
+Deno.test("take", () => {
+  assertEquals(take.parse(""), {
+    success: false,
+    message: parseErrors.takeError,
+    position: { line: 1, column: 0 },
+  });
+
+  assertEquals(take.parse("monad"), {
+    success: true,
+    results: [{
+      value: "m",
+      remaining: "onad",
+      position: { line: 1, column: 1 },
+    }],
+  });
+});
 
 Deno.test("letter", () => {
   assertEquals(letter.parse("m"), {
