+++ src/common/characters.ts
@@ -1,6 +1,48 @@
+import { regex } from "$common";
 import { parseErrors } from "../errors.ts";
-import type { Parser } from "../index.ts";
-import { regex } from "./mod.ts";
+import { createParser, type Parser } from "../parser.ts";
+import { updatePosition } from "../utils.ts";
+
+/**
+ * Consumes the next character of the input and fails if the input is empty.
+ *
+ * @example Non-empty input
+ *
+ * ```ts
+ * const { results } = take.parse("hello");
+ * // [{value: 'h', remaining: 'ello', ...}]
+ * ```
+ *
+ * @example Empty input
+ *
+ * ```ts
+ * const { message } = take.parse("");
+ * // "Unexpected end of input"
+ * ```
+ */
+export const take: Parser<string> = createParser(
+  (input, position) => {
+    if (input.length === 0) {
+      return {
+        success: false,
+        message: parseErrors.takeError,
+        position,
+      };
+    }
+
+    const consumed = input[0];
+    const newPosition = updatePosition(position, consumed);
+
+    return {
+      success: true,
+      results: [{
+        value: consumed,
+        position: newPosition,
+        remaining: input.slice(1),
+      }],
+    };
+  },
+);
 
 /**
  * Parses alphanumeric characters (0+)
