+++ src/common/mod.ts
@@ -1,5 +1,5 @@
 export * from "./list.ts";
-export { regex } from "./regex.ts";
+export * from "./regex.ts";
 export * from "./tokens.ts";
 export * from "./whitespaces.ts";
 export * from "./numbers.ts";
