+++ src/utils.ts
@@ -1,0 +1,27 @@
+import type { Position } from "./types.ts";
+
+/**
+ * Compute the new position from the current position and the consumed string
+ */
+export const updatePosition = (
+  position: Position,
+  consumed: string,
+): Position => {
+  const lines = consumed.split("\n");
+  const newLines = lines.length > 1;
+
+  return {
+    line: position.line + lines.length - 1,
+    column: newLines
+      ? lines.at(-1)!.length
+      : position.column + lines.at(-1)!.length,
+  };
+};
+
+/**
+ * Sort positions in a descending (line, column) order
+ */
+export const sortPosition = (a: Position, b: Position): number => {
+  if (a.line !== b.line) return b.line - a.line;
+  return b.column - a.column;
+};
